// frontend/src/lib/api/models/AssessmentScopeResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AssessmentScopeResponse = {
    id: string;
    moduleId: string;
    name: string;
    coveredWeeks: Array<number>;
    status: string;
    createdAt: string;
    updatedAt: string;
};